Reorganise test file structure
